+++ tests/unit/test_gama.py
@@ -14,20 +14,18 @@ def test_reproducible_initialization():
 
 
 def test_gama_fail_on_invalid_hyperparameter_values():
-    # `delete_cache` is only called when the initialization failed to raise a ValueError.
-    # In that case, we would need to delete the created cache directory. Otherwise, it would not even be created.
     with pytest.raises(ValueError) as e:
-        gama.GamaClassifier(max_total_time=0).delete_cache()
+        gama.GamaClassifier(max_total_time=0)
     assert "max_total_time should be integer greater than zero" in str(e.value)
 
     with pytest.raises(ValueError) as e:
-        gama.GamaClassifier(max_total_time=None).delete_cache()
+        gama.GamaClassifier(max_total_time=None)
     assert "max_total_time should be integer greater than zero" in str(e.value)
 
     with pytest.raises(ValueError) as e:
-        gama.GamaClassifier(max_eval_time=0).delete_cache()
+        gama.GamaClassifier(max_eval_time=0)
     assert "max_eval_time should be None or integer greater than zero" in str(e.value)
 
     with pytest.raises(ValueError) as e:
-        gama.GamaClassifier(n_jobs=-2).delete_cache()
+        gama.GamaClassifier(n_jobs=-2)
     assert "n_jobs should be -1 or positive integer" in str(e.value)