+++ docs/source/releases.rst
@@ -5,9 +5,14 @@ Version 20.1.0
 --------------
 Features:
  - Encoding of ARFF files may now be specified with the `encoding` parameter in {fit/predict/score}_arff calls.
+ - Set `max_pipeline_length` on initialization to impose a maximum number of steps in your pipelines.
 
 Bugfixes:
  - Reading ARFF markers (such as @data and @attribute) is now correctly case insensitive.
+
+Maintenance:
+ - Evaluation results are no longer saved to disk but kept in memory only.
+   Consequently the `cache_dir` hyperparameter has been removed.
 
 Changes:
  - Pipelines fitted during search are now used in the ensemble, instead of retraining the pipeline.